PEP 772: Open up Packaging Council elections widely
Rename "voting members" to "College of Packaging Electors," and "member"
to "Elector" etc., to avoid any impression that being part (or not part)
of this body is reflective of membership or welcome in the Python
packaging community itself.

Open up the College to all PSF voting members. Make its membership
public. Require a supermajority of the Packaging Council to re-add a
member who was removed by such a supermajority vote.

Explicitly advertise that the requirements of PSF contributing
membership are likely to be met by people who are involved as
contributors or engaged users in any Python-related project, even those
outside the PyPA aegis.

Explicitly state the expectations on Electors, which are to receive
emails about votes approximately once a year and to vote. Remove the
implicit requirement that Electors have a discussion forum among
themselves.

Make the mechanism for joining the College or proposing a
vote of no confidence a button in the PSF member web UI, given the lack
of other forum.

Given the widened voter base, require multiple seconds for a vote of no
confidence, and state that quorum is all Electors, not just those who
request a ballot. Also, rate-limit no confidence votes.

Remove the conflict-of-interest exception for PSF members, as we are
encouraging everyone to become PSF members.

Tighten the definition of "significant affiliation".

Ban Steering Council members from Packaging Council membership, and
explicitly give the Steering Council authority to change this document
in the ordinary process (PEP 1 leaves open the process for amending
PEPs).

Clarify the anonymity provisions around ballots and create a process for
deanonymizing ballots.